feat: emoji -> icon
